% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\docType{methods}
\name{piieffect}
\alias{piieffect}
\alias{piieffect,character-method}
\title{piieffect}
\usage{
piieffect(value)

\S4method{piieffect}{character}(value)
}
\arguments{
\item{value}{A character that will be returned}
}
\description{
Long description for function goes here
and here...
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Generic/Function
}}

\examples{

output <- piieffect("hello")

}
\author{
Isabel Fulcher
}
