% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\docType{methods}
\name{piieffect}
\alias{piieffect}
\alias{piieffect,data.frame,character,character,character,vector,vector,vector,numeric,numeric-method}
\title{piieffect}
\usage{
piieffect(data, outcome, intermediate, exposure, covariates.outcome,
  covariates.intermediate, covariates.exposure, interaction, astar)


  \S4method{piieffect}{data.frame,character,character,character,vector,vector,vector,numeric,numeric}(data,
  outcome, intermediate, exposure, covariates.outcome = c(1),
  covariates.intermediate = c(1), covariates.exposure = c(1),
  interaction = 1, astar = 0)
}
\arguments{
\item{data}{A dataframe}

\item{outcome}{The variable name for the outcome variable in data}

\item{intermediate}{The variable name for the intermediate variable in data}

\item{exposure}{The variable name for the exposure variable in data}

\item{covariates.outcome}{A vector of variable names for covariates to be included in the outcome model}

\item{covariates.intermediate}{A vector of variable names for covariates to be included in the outcome model}

\item{covariates.exposure}{A vector of variable names for covariates to be included in the exposure model}

\item{interaction}{A binary variable indicating if an interaction term between intermediate and exposure is needed}

\item{astar}{A numeric value for the level of the exposure wanted for comparison}
}
\description{
Long description for function goes here
and here...
}
\section{Methods (by class)}{
\itemize{
\item \code{data = data.frame,outcome = character,intermediate = character,exposure = character,covariates.outcome = vector,covariates.intermediate = vector,covariates.exposure = vector,interaction = numeric,astar = numeric}: Generic/Function
}}

\examples{

simdata <- readRDS(system.file("rds","simdata1.rds",package="frontdoorpiie"))
simdata$c1c2 <- simdata$c1*simdata$c2
output <- piieffect(data=simdata,outcome="y",intermediate="m",exposure="a",covariates.outcome=1,covariates.intermediate=c("c1"),covariates.exposure=c("c1","c2","c1c2"),interaction=1,astar=0)

}
\author{
Isabel Fulcher
}
